pyproject.toml
[project]
name = "pdf-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "chromadb>=1.0.15",
    "click==8.2.1",
    "colorama==0.4.6",
    "fastapi==0.115.13",
    "h11==0.16.0",
    "idna==3.10",
    "nltk~=3.9.1",
    "openai~=1.90.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pymupdf==1.23.8",
    "python-dotenv==1.1.0",
    "python-multipart>=0.0.20",
    "requests~=2.32.4",
    "sentence-transformers~=4.1.0",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "streamlit~=1.37.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "uvicorn==0.34.3",
]
